Update monitor tests
Update tests for statistics requests using abinary protocol. Now test check requests with all possible varitions of categories and backends.
Add tests for http statistics requests.
